qca-wifi: Add regulatory API for getting AP chanlist
Add a regulatory API that will retrieve the one of
the following channel list for AP mode:
1. Current channel list
2. Master channel list with SP power
3. Master channel list with LPI power
4. Master channel list with VLP power

For SP power, the API will intersect the SP channel
list with the AFC master channel list in the
regulatory.

Additionally, intersect the channel list with the
active frequency range of the hardware as per the
regulatory rules.

CRs-Fixed: 3013394
Change-Id: I52bb933c87cc8fc77f487d59c1104216329560b0