qca-wifi: Clear the MBSSID_RESTART flag if vdev chan config is not valid
In case of CSA without serialisation, there is a possibillity that vdevs
are brought down even before restart is triggered in vdev sm.
Now if a new vdev is brought up and is going for individual restart,
because of this pdev flag it will consider it MVR leading to MVR
timeout.
The MBSSID_RESTART flag is cleared in this case if none of the vdev is
active. 
But if CSA is on DFS channel, then vdevs will go via suspend
restart instead of CSA restart. 
If this CSA is triggered on repeater, as part of csa_complete, STA vdev
will be brought down. Then as part of sta vdev going down,vdev state check
for active vdevs will return false and the MBSSID_RESTART bit
will be reset. If any of the vdev has moved to restart state
before this flag is reset will get stuck in restart state while 
other vdevs will go for individual restart. The vdev stuck in
restart state will lead to MVR timeout
Instead check for vdev not to be in init and stop state
to include all valid active states

Change-Id: I07c0567c1d6d199967a096ace69074ff544751a5
CRs-Fixed: 3055753