qca-wifi: Preserve autoswitch_desmode in MBSSID
Bring up 2 AP VAPS with interCACChan as 5180MHZ (IEEE:36), preCACEn
set to 1, htmode as HT160 and desired channel as 100. In case of
HK, the agile detector can support a max width of 80MHZ. Hence to do
preCAC on 100 HT160, we have to do preCAC twice, once on each HT80
(5530 and 5610). dfs_autoswitch_des_mode is set to HT160 and
dfs_autoswitch_chan is set to 100HT160, agile synthesizer begins preCAC
on first HT80 of channel 100 (a.k.a 5530MHZ).

In the meantime, interCAC chan is set to 36HT80 and
the vap desmode is also changed to HT80. When the second vap tries to
do a vdev start, since desmode is altered internally to HT80, it checks
if 100 HT80 is preCAC done and if not sets "dfs_autoswitch_des_mode" to HT80,
dfs_autoswitch_chan to 100HT80 and comes up on interCAC channel (36HT80).

Hence after preCAC is completed on 5530MHZ, we switch to 100 HT80
mode since the second vap had altered the user desired
dfs_autoswitch_chan and dfs_autoswitch_des_mode from HT160 to HT80.

Preserve the user set dfs_autoswitch_des_mode in multi-vap scenerio.
Do not allow dfs_autoswitch_des_mode to change to interCAC chan's des_mode.

CRs-Fixed: 3062044
Change-Id: Ifcdd138f48afeb4e2b20a0929289c7ce2602e6c6