qca-wifi: Use the n_valid subchannel count in preCAC done check
While bringing up an AP in channel 100 HT20, even is preCAC is disabled,
channel 100 HT20 is marked as preCAC done and CAC is skipped.

While checking if preCAC is done for a particular
frequency, the tree nodes are traversed considering each node
has subchannel count in powers of 2. While traversing from HT160 (root)
to HT20, the subchannel count is 8->4->2->1. When we reach
100HT20, the number of subchannels is 1 and only if
CACed subchannels count is also 1, 100HT20 is considered preCAC done.
With the addition of 240MHz, the subchannel count is no longer
in powers of 2. While traversing from 240MHz, the subchannel count,
if considered in powers of 2, is 12->6->3->1->0 (instead of
12->8->4->2->1). While checking for 100HT20, the number of
subchannels is 0 and since CAC subchannels count is also 0, 100HT20
is always considered CAC done.

While checking if preCAC is done for a particular center frequency,
use the n_valid subchannel count for every child node traversed
by the API, instead of dividing the initial n_valid subchannel
count by 2.