qca-wifi: Update rate entries in table using rate and rate index combo
As per the current implementation, peer rate extended stats are updated
considering the rate index as the key value.
1 rate index can correspond to different rate values as per the guard
interval duration and in such scenario, the rate value is updated with
latest rate value in the entry corresponding to that rate index.

For correct and valid output, consider the combination of rate and
rate index while updating the entries in the table. With this change,
in flush output, the combination of rate and rate index is unique.

Change-Id: I0f4094ac4273a7fd6d5c5eb3bec0d310c79a02c1
CRs-Fixed: 3161169