qca-wifi: Check for ETSI domain before enabling preCAC
preCAC is a feature specific to ETSI countries.
However, when country is switched from DK to US dynamically, preCAC
is initiated. This is because dfs_is_agile_precac_enabled() API
fails to check for the current regdomain.

Add regdomain check in dfs_is_agile_precac_enabled() so that preCAC
begins only on ETSI countries.

CRs-Fixed: 3166878
Change-Id: Iff7a9a5f4801e8c8e156655b99ee5740adb576a4