qca-wifi: Support NOL IE and ADFS for sta in 5 and 10MHz
NOL IE contains a bitmap to indicate the affected subchannels (each of
20MHz BW) and the first subfrequency. With the addition of 5 and 10MHz
DFS channels, NOL IE should be able to contain the affected 5/10MHz
channels.

With the introduction of a range based system to maintain NOL entries,
the NOL IE should be appropriately changed. Modify the NOL IE to
share the affected ranges using variable BW bitmap.

Also add support for ADFS in client mode in 5/10MHz and the ADFS
completion bit from stations for 5/10MHz in the new 5/10MHz preCAC
list.

Change-Id: If36ca69f57c274362eacc7092f0796b4098d7ac8