qcacmn: Recalculate start ptr of beacon buf before appending ssid
When hidden SSID is enabled, the offset at which the SSID IE exists
in a beacon buffer is first calculated, after which the extra length
is added to the beacon skb.
Since this increase in length may change the start of the data in
skb, reassign the start pointer remembered locally
to the data pointer of the skb before adding the SSID configured.

Change-Id: Ida32927de5986105efdc10c5115466fe071aa9f3