qca-wifi: Use rssi_chain value in uint8_t format for EWMA averaging
The value of rssi_chain can vary between -128 to 127.

For peer extended rate stats, average rssi_chain is computed using
EWMA averaging mechanism which expects the variable data type to be
of unsigned long format.

In case of negative rssi_chain value, avg value turns out to be a
very big value which does not hold any significance.

In order to fix this, use rssi_chain in uint8_t format for average
computation and in the flush output, print the average value in
int8_t format.

Change-Id: I5bdec86ecd722d59d3e2641d11ccd875b368818a
CRs-Fixed: 3177194