qca-wifi: Extract the Quiet duration value for tbtt count zero
Host extracts the Quiet information from FW as part of SWBA event.
The Quiet duration is not extracted when the tbtt count is zero.

When MLD AP receives the MLO Assoc request during the Quiet duration,
it is expected to include the Quiet IE in the MLO Assoc response
to indicate when the Quiet triggered AP will be available.
In the current code, Quiet IE will not be added as Quiet duration is
set to zero after announcing the Quiet IE in the beacon.

To fix this issue, extract the Quiet IE related information from
the SWBA event even when the tbtt count is zero.

Change-Id: I079deba4c0656e088c4347140ab2b121d4286f1f