qca-wifi-oss: BW Expansion for Rolling CAC
When BW Expand feature is enabled, RCAC picks the
user configured channel as next RCAC channel instead
of Random Channel Selection approach. Any subchannels
that are present in the user configured channel is
free of NOL, CAC not completed and not a current channel,
then the channel is selected as next RCAC channel and start
running CAC on that subchannel.

Once Rolling CAC is completed on the selected channel,
perform a seamless channel change to the RCAC completed
channel. Check if there is a possibility of bandwidth
expansion. If it is possible, move to higher bandwidth
or choose the next RCAC channel based on the BW Expand
algorithm.

CRs-Fixed: 3230896
Change-Id: I30c62fffa4a1f30a4ce2504b2a5550bdd98de134