qca-wifi: Add callback for SCS peer lookup and rule match
SCS rules are populated into ECM database.

On rule match for first packet in the flow, ECM shall query the
host driver using the registered callback by passing
destination mac address and matching rule id as arguments.

Host driver shall then look for the peer using mac address
and check if that rule is applicable for that peer or not.

Based on the result returned by host driver, ECM shall decide
to proceed or not with traffic prioritization.

Change-Id: I608ee835e4c8b650294e80209493126e675877dc
CRs-Fixed: 3228764