qcawifioss: Run preCAC and rCAC HRtimers in softIRQ context
1. The preCAC timeout callback function is divided into two parts: HardIRQ
and workqueue. If preCAC expires, the callback function runs in a HardIRQ
context and the rest of the task runs in a workqueue to avoid deadlock
issues.
2. rCAC timer uses normal timer instead of HRtimer

Issues:
1. preCAC Workqueue tasks run in a low-level context. The task's
execution might have been delayed.
2. rCAC timer uses normal timer instead of HRtimer. In the 5.4 kernel, a
rCAC timer expiry is delayed at the rate of 1 sec per minute (approx).

To fix this, the following changes are needed:
1. Use QDF_CONTEXT_TASKLET instead of QDF_CONTEXT_HARDWARE as a argument in
qdf_hrtimer_init. If it is a "QDF_CONTEXT_TASKLET," the HR timer
initialises with softIRQ mode(HRTIMER_MODE_SOFT) and runs in softRQ
context when timer expires.

2. To avoid the delay, use the HR(High Resolution) timer for rCAC and
follow step 1.

Change-Id: Iabe68adfb131a38dcb097783d9132bcd6a7fd1da
CRs-Fixed: 3239452