qca-wifi: Use host queue id for SAWF stats
For high bandwidth traffic, initial packets are not added to the
ingress packet count as host needs remapped TID and queue id from
the FW to populate the stats. Use host queue id for storing the
stats, thereby eliminating host's dependency on the FW event.
The stats can be translated into per TID per queue as needed based
on the available mapping.

Change-Id: I4a5f6458e0270095437780fff0d310dc6f4c04d8
CRs-Fixed: 3242979