qca-wifi-oss: Expand Bandwidth if higher bandwidth available
When Bandwidth Expansion feature is enabled, when preCAC is
completed on the channel, the agile SM handles the case
DFS_AGILE_SM_EV_AGILE_DONE. Check if there is a possibility
of expanding the current agile bandwidth to higher bandwidths.
The channel to which its trying to expand the bandwidth must
be subset of the user configured channel. If the condition
is satisfied, do channel change with higher bandwidth.

In case of Rolling CAC, when CAC is completed on RCAC channel,
move to the RCAC completed channel and then check the
possibility of expanding to higher bandwidths.

CRs-Fixed: 3229895
Change-Id: Idaa233f7067c8a581298584759595acafcdb356f