qcacld-3.0: Add one to max peer per vdev for STA roaming
When roaming in progress and BMISS happen, it is possible peer attach
vdev fail, because old bss peer not deleted by additional reference
temporary.

If enable TDLS, this issue is mitigated because max peer number is
greater than 2 if add TDLS max peer number. But on some project which
disable TDLS, it is possible encounter this race condition.

Change-Id: I610f966df586bbe850dde22820fb48aaafc8d307
CRs-Fixed: 3236737