qca-wifi-oss: Do not access invalid channel enum entry
channel enum is used as an array index in regualtory's channel list
array (pdev_priv_obj->cur_chan_list). This chan_enum when computed
for a secondary 40 MHz channel can be invalid and when
INVALID_CHANNEL is used as an array index, it causes out of band
call trace in KASAN.

Check for INVALID_CHANNEL enum before accessing the array index.

CRs-Fixed: 3244906
Change-Id: Ifc125091b36ca0c27ef8864e327e7ae5bd7c4ed7