qca-wifi-oss: Check preCAC done for all channels
In BW Expansion feature, When user configured channel
is included NOL list after radar found, agileCAC starts
running on other available channels. When NOL expiry
event is received, the current agileCAC running on any
other channel should abort and start CAC in user configured
channel.

In case of chan 36 HT160, after the channel comes out of
NOL list, the agileCAC is not happening immediately, the
agileCAC completes on the other channel and then start
CAC on user configured channel. The reason for the behavior
is chan 36 HT160 is mixed of NON-DFS and DFS channel. The
condition for agileCAC to start on user configured channel
is it should not be CAC done already. In this case, the
NON DFS channels represent CAC already completed, therefore
it is not eligible for preCAC.

Whenever checking the eligibility of preCAC, check if all
the subchannels are CAC completed. If all subchannels are
CAC completed, we can skip preCAC on that user configured
channel.

CRs-Fixed: 3268422
Change-Id: Ie29708e55b09eba2c1323bfaf6664831ac231f9e