qca-wifi-oss: Add RCSA support for 320 MHz BW
dfs_nol_ie_bitmap is a 8 bit value representing the bitmap of
radar infected sub-channels of the operating channel.

dfs_nol_ie_startfreq is always set to the first 20 MHz sub-channel
(lefmost) of the operating channel BW. A radar affected 320 MHz channel
(240 - 80) cannot fit into this architecture as there are 12 subchannels
to be represented from the leftmost frequency.

In order to accommodate the 320 MHz radar handling and also to ensure
backward compatibility, retain the same 8 bit representation of the
dfs_nol_ie_bitmap and let dfs_nol_ie_startfreq to indicate the first
radar infected 20 MHz subchannel instead of the first sub-channel of the
operating BW.
Since the max number of radar infected sub-channels can be 3 (if it is
a chirp) and these channels will also be contiguous, use the
dfs_nol_ie_bitmap to indicate the successive radar channels from the
dfs_nol_ie_startfreq.

No changes are needed in the parsing logic as it already computes the
NOL channels based on dfs_nol_ie_startfreq.

CRs-Fixed: 3271982
Change-Id: Ic9ce84fd720fb7c5018d4b89c31a27d967674a7e