qcacld-3.0: Add pwrmode argument to callers
The current channel list alone is not enough to represent the
capability of the chip or device. Given a channel, in many
cases it may be required to know all the power modes that are
supported by this channel.

Add pwrmode argument to callers of wlan_reg_get_opclass_details.

Change-Id: Ia470ae99cb374e4cfdbea8ce77cceabba5c7c959
CRs-Fixed: 3357155