qcacmn: qdf: Remove obsolete suppressed logging macros
The original implementation of suppressed logging used identifiers
with a misspelling of "suppressed."

Change [1] renamed the identifiers, but for 2 public functions it
defined macros for the original names to keep from breaking existing
clients.

Change [2] updated the only existing client to use the new names.

Now that the old names are no longer used, remove the macros.

[1] Change-Id I0107b157374046ab8f3467801563c02f0f2651fa
    qcacmn: qdf: Rename misspelled "suppressed" variables/functions

[2] Change-Id I4fcd506b39cedd6f094f93ee900cf9ce8be759c1
    qcacmn: Use renamed QDF suppressed logging APIs

Change-Id: I78f3b245b355ba937b15fdfa5b4834fb58a62a62
CRs-Fixed: 3313687