qca-wifi: Change Agile Detector id for 11BE
In case of 11BE, the detector id for agile channel is 1 and
for home channel the detector id is 0

In bangradar, channel width for agile channel is determined and the
freq offset check is validated based on the channel width for 11BE.

A new capability bit from FW is introduced that provides
ADFS 320MHz support information. supports_adfs_320 signifies
chainmask capability service bit if FW supports ADFS on 320 MHz
channel width. Add this capability bit(supports_adfs_320) in host.

Change-Id: Ifc9d93cc0e1f97e44267922cdfd491547d133e93