qcacld-3.0: Reduce bw if selected non-DFS channel has DFS bonding
policy_mgr_get_nondfs_preferred_channel() return a non-DFS channel
with channel bw 20 MHz, but if the original SAP bw is 160 MHz and
channel is on 5 GHz, it always has DFS bonding channel.

So reduce channel bw if detect the bonding channel is DFS channel.
It can make STA connect to 5 GHz ref-AP even SAP is on 5 GHz channel
with 160 MHz bw.

Change-Id: I842d9dfb1070b900e52841b4ec5372af8550ad8e
CRs-Fixed: 3318696